chore: rollback dedicated page for add/edit, remove dialog
