+++ src/pages/events/edit-event-category-group-page.js
@@ -11,7 +11,7 @@
  * limitations under the License.
  * */
 
-import React from "react";
+import React, { useEffect } from "react";
 import { connect } from "react-redux";
 import T from "i18n-react/dist/i18n-react";
 import { Breadcrumb } from "react-breadcrumbs";
@@ -26,74 +26,70 @@ import {
   removeAllowedGroupFromGroup,
   getEventCategoryGroupMeta
 } from "../../actions/event-category-actions";
-import AddNewButton from "../../components/buttons/add-new-button";
 
-// import '../../styles/edit-summit-attendee-page.less';
+const EditEventCategoryGroupPage = ({
+  currentSummit,
+  entity,
+  allClasses,
+  match,
+  history,
+  getEventCategoryGroup,
+  resetEventCategoryGroupForm,
+  saveEventCategoryGroup,
+  addCategoryToGroup,
+  removeCategoryFromGroup,
+  addAllowedGroupToGroup,
+  removeAllowedGroupFromGroup,
+  getEventCategoryGroupMeta
+}) => {
+  const groupId = match.params.group_id;
 
-class EditEventCategoryGroupPage extends React.Component {
-  componentDidMount() {
-    const { allClasses } = this.props;
-    const groupId = this.props.match.params.group_id;
-
+  useEffect(() => {
     if (!groupId) {
-      this.props.resetEventCategoryGroupForm();
+      resetEventCategoryGroupForm();
     } else {
-      this.props.getEventCategoryGroup(groupId);
+      getEventCategoryGroup(groupId);
     }
+  }, [groupId]);
 
+  useEffect(() => {
     if (allClasses.length === 0) {
-      this.props.getEventCategoryGroupMeta();
+      getEventCategoryGroupMeta();
     }
-  }
+  }, [allClasses.length]);
 
-  componentDidUpdate(prevProps, prevState, snapshot) {
-    const oldId = prevProps.match.params.group_id;
-    const newId = this.props.match.params.group_id;
+  const handleSubmit = (values) =>
+    saveEventCategoryGroup(values).then(() => {
+      history.push(`/app/summits/${currentSummit.id}/event-category-groups`);
+    });
 
-    if (oldId !== newId) {
-      if (!newId) {
-        this.props.resetEventCategoryGroupForm();
-      } else {
-        this.props.getEventCategoryGroup(newId);
-      }
-    }
-  }
+  const title = entity.id
+    ? T.translate("general.edit")
+    : T.translate("general.add");
+  const breadcrumb = entity.id ? entity.name : T.translate("general.new");
 
-  render() {
-    const { currentSummit, entity, allClasses, errors, match } = this.props;
-    const title = entity.id
-      ? T.translate("general.edit")
-      : T.translate("general.add");
-    const breadcrumb = entity.id ? entity.name : T.translate("general.new");
+  if (!allClasses.length || !currentSummit) return <div />;
 
-    if (!allClasses.length) return <div />;
-
-    return (
-      <div className="container">
-        <Breadcrumb data={{ title: breadcrumb, pathname: match.url }} />
-        <h3>
-          {title}{" "}
-          {T.translate("edit_event_category_group.event_category_group")}
-          <AddNewButton entity={entity} />
-        </h3>
-        <hr />
-        {currentSummit && (
-          <EventCategoryGroupForm
-            currentSummit={currentSummit}
-            allClasses={allClasses}
-            entity={entity}
-            errors={errors}
-            onTrackLink={this.props.addCategoryToGroup}
-            onTrackUnLink={this.props.removeCategoryFromGroup}
-            onAllowedGroupLink={this.props.addAllowedGroupToGroup}
-            onAllowedGroupUnLink={this.props.removeAllowedGroupFromGroup}
-            onSubmit={this.props.saveEventCategoryGroup}
-          />
-        )}
-      </div>
-    );
-  }
-}
+  return (
+    <div className="container">
+      <Breadcrumb data={{ title: breadcrumb, pathname: match.url }} />
+      <h3>
+        {title} {T.translate("edit_event_category_group.event_category_group")}
+      </h3>
+      <hr />
+      <EventCategoryGroupForm
+        currentSummit={currentSummit}
+        allClasses={allClasses}
+        entity={entity}
+        onSubmit={handleSubmit}
+        onTrackLink={addCategoryToGroup}
+        onTrackUnLink={removeCategoryFromGroup}
+        onAllowedGroupLink={addAllowedGroupToGroup}
+        onAllowedGroupUnLink={removeAllowedGroupFromGroup}
+      />
+    </div>
+  );
+};
 
 const mapStateToProps = ({
   currentSummitState,
